+++ test/e2e/steps/tap_search_result.dart
@@ -51,24 +51,24 @@ final StepDefinitionGeneric tapUserInSearchResults =
             final finder = context.world.appDriver
                 .findByKeySkipOffstage('SearchContact_$id');
 
-            if (await context.world.appDriver.isPresent(finder)) {
-              await context.world.appDriver.scrollUntilVisible(
-                finder,
-                scrollable: find.descendant(
-                  of: find.byKey(const Key('SearchScrollable')),
-                  matching: find.byWidgetPredicate((widget) {
-                    // TODO: Find a proper way to differentiate [Scrollable]s from
-                    //       [TextField]s:
-                    //       https://github.com/flutter/flutter/issues/76981
-                    if (widget is Scrollable) {
-                      return widget.restorationId == null;
-                    }
-                    return false;
-                  }),
-                ),
-                dy: 100,
-              );
+            await context.world.appDriver.scrollUntilVisible(
+              finder,
+              scrollable: find.descendant(
+                of: find.byKey(const Key('SearchScrollable')),
+                matching: find.byWidgetPredicate((widget) {
+                  // TODO: Find a proper way to differentiate [Scrollable]s from
+                  //       [TextField]s:
+                  //       https://github.com/flutter/flutter/issues/76981
+                  if (widget is Scrollable) {
+                    return widget.restorationId == null;
+                  }
+                  return false;
+                }),
+              ),
+              dy: 100,
+            );
 
+            if (await context.world.appDriver.isPresent(finder)) {
               await context.world.appDriver.scrollIntoView(finder);
               await context.world.appDriver.tap(
                 finder,
@@ -86,24 +86,24 @@ final StepDefinitionGeneric tapUserInSearchResults =
             final finder = context.world.appDriver
                 .findByKeySkipOffstage('SearchUser_$userId');
 
-            if (await context.world.appDriver.isPresent(finder)) {
-              await context.world.appDriver.scrollUntilVisible(
-                finder,
-                scrollable: find.descendant(
-                  of: find.byKey(const Key('SearchScrollable')),
-                  matching: find.byWidgetPredicate((widget) {
-                    // TODO: Find a proper way to differentiate [Scrollable]s from
-                    //       [TextField]s:
-                    //       https://github.com/flutter/flutter/issues/76981
-                    if (widget is Scrollable) {
-                      return widget.restorationId == null;
-                    }
-                    return false;
-                  }),
-                ),
-                dy: 100,
-              );
+            await context.world.appDriver.scrollUntilVisible(
+              finder,
+              scrollable: find.descendant(
+                of: find.byKey(const Key('SearchScrollable')),
+                matching: find.byWidgetPredicate((widget) {
+                  // TODO: Find a proper way to differentiate [Scrollable]s from
+                  //       [TextField]s:
+                  //       https://github.com/flutter/flutter/issues/76981
+                  if (widget is Scrollable) {
+                    return widget.restorationId == null;
+                  }
+                  return false;
+                }),
+              ),
+              dy: 100,
+            );
 
+            if (await context.world.appDriver.isPresent(finder)) {
               await context.world.appDriver.scrollIntoView(finder);
               await context.world.appDriver.tap(
                 finder,
